services/auth-service/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
pydantic[email]
pydantic-settings
passlib[bcrypt]
bcrypt==4.0.1
PyJWT
pyotp
redis[hiredis]
